refactor : BottomSheet > useModal 사용하여 리팩토링
