--- src/app/(main)/grouping/create/components/inputSection/ImageSection.client.tsx
@@ -1,17 +1,0 @@
-'use client';
-import { useCreateGroupContext } from '../CreateGroupContext';
-import ImageFrame from '@/components/common/ImageFrame';
-import { useRef } from 'react';
-
-export default function ImageSection() {
-  const imgRef = useRef<HTMLInputElement>(null);
-  const { setValue } = useCreateGroupContext();
-
-  return (
-    <ImageFrame
-      setImageUrl={(value: string) => setValue('imageUrl', value)}
-      shape="square"
-      ref={imgRef}
-    />
-  );
-}
